power-energy-calculations/historical-data/historical_calculation/requirements.txt
boto3==1.26.163
botocore==1.29.163
jmespath==1.0.1
joblib==1.3.0
mysql-connector-python==8.0.33
numpy==1.24.0
pandas==2.0.3
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1   
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.16
